% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the locations of accidents for a specific state in a year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A value that indicates a state ID. Must be able to be casted
to an integer between 1 and 56. Error will occur if \code{year} is not a single,
valid value or there is no existing data for that year; or if \code{state.num}
is not a valid state number.}

\item{year}{A value that specifies a year. Must be able to be casted to an
integer that specifies a valid year.}
}
\value{
None. Draws a map a of a state and points indicating locations of
accidents, if there are accidents to plot. Otherwise draw nothing.
}
\description{
Plots the locations of accidents for a specific state in a year.
}
\examples{
\dontrun{
fars_map_state(1, 2013)
fars_map_state(19, 2014)
# error will be produced:
fars_map_state(100, 2013) # invalid state.num
fars_map_state(1, 2017) # invalid year
}
}
