% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data from a csv file and constructs a tibble.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character specifying the csv file to read. A non-existing
filename will produce an error.}
}
\value{
a tibble containing the csv data.
}
\description{
Read data from a csv file and constructs a tibble.
}
\examples{
\dontrun{
fn <- "data/accident_2013.csv.bz2"
data <- fars_read(fn)
}
# filename that does not exist gives an error.
\dontrun{
fn <- NULL
data <- fars_read(fn)
}
}
