% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Returns occurrence of accidents in a specific year.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector (atomic or list) of or a single numeric or character
specifying years of accidents to return. Giving \code{years} that cannot
be casted to a numeric, or for which no data exist will return NULL and
a warning message.}
}
\value{
A tibble with 'MONTH' and 'year' columns specifying the month of
accidents in the year.
}
\description{
Returns occurrence of accidents in a specific year.
}
\examples{
dat <- fars_read_years(2013)
dat <- fars_read_year(c(2013, 2014))
dat <- fars_read_year("2013")
\dontrun{
dat <- fars_read_year(c(2013, "hello")) # returns NULL for "hello"
}
}
