% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes the number of accidents in each month for specific years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector (atomic or list) of or a single numeric or character
specifying years of accidents to return. \code{years} input for which no
data exist or that cannot be casted to a year will be skipped. An error
will occur when no valid inputs are provided.}
}
\value{
A tibble that summarizes that number of accidents in the years
for which data exist.
}
\description{
Summarizes the number of accidents in each month for specific years.
}
\examples{
\dontrun{
summary_table <- fars_summarize_years(2013)
summary_table <- fars_summarize_years(c(2013, 2014))
summary_table <- fars_summarize_years("2013")

# no column will be returned for invalid inputs
summary_table <- fars_summarize_years(c(2013, "hello"))
# error for single invalid inputs
summary_table <- fars_summarize_years("hello")
}
}
