% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generates a filename for a specific year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character or numeric specifying a year. Input \code{year} that
cannot be casted to an integer will result in an error or an empty return array.}
}
\value{
A string in the format "accidient_\%d.csv.bz2".
}
\description{
Generates a filename for a specific year.
}
\examples{
fn <- make_filename(year="2020")  
fn <- make_filename(year=2020)
# input values that cannot be casted to an integer will result in an error or 
# empty return array 
\dontrun{
fn <- make_filename(year=NULL)
fn <- make_filename(year="hello")
}
}
